added check check if the $_GET[id] is not empty

+++ delete/delete.php
@@ -4,11 +4,17 @@ $db-> setAttribute(PDO::ATTR_DEFAULT_FETCH_MODE, PDO::FETCH_ASSOC);
 ?>
 <link rel="stylesheet" type="text/css" href="../add_new/add_proj.css"><link href="https://fonts.googleapis.com/css?family=Roboto+Condensed" rel="stylesheet">
 <?php
-$deleteProject = 'DELETE FROM `projects` WHERE `id` = ?;';
-$query = $db->prepare($deleteProject);
-$query->execute([$_GET['id']]);
-$project = $query->fetch();
+if (!empty($_GET['id'])) {
+    $deleteProject = 'DELETE FROM `projects` WHERE `id` = ?;';
+    $query = $db->prepare($deleteProject);
+    $query->execute([$_GET['id']]);
+    $project = $query->fetch();
+    echo '<h1>This project has been deleted from the Database</h1>
+    <a class="button back" href="../admin_page/index.php">Back to Admin Panel</a>';
+} else {
+    echo '<h1>Unable to delete this project</h1><h3>Please check the details of this project in the database</h3>
+    <a class="button back" href="../admin_page/index.php">Back to Admin Panel</a>';
+}
 ?>
-<h1>This project has been deleted from the Database</h1>
-<a class="button back" href="../admin_page/index.php">Back to Admin Panel</a>
+
